requirements.txt
torch~=2.5.0
matplotlib~=3.7.1
pandas~=2.2.3
torchvision~=0.20.0
pillow~=11.0.0
numpy~=1.24.2
tqdm~=4.67.0
scikit-image~=0.24.0